requirements.txt
streamlit>=1.28.0
yfinance>=0.2.31
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
pypfopt>=1.5.5
feedparser>=6.0.10
requests>=2.31.0
finnhub-python>=2.4.18
python-dotenv>=1.0.0
alpaca-py>=0.20.0
